refactor: move download2mp3.js into src/download/
